+++ RTU_Master/RTU_Master/include/main.h
@@ -18,4 +18,4 @@ using namespace std;
 
 
 
-uint16_t crc16table(uint8_t* ptr, uint16_t len);
+uint16_t crc16table(uint16_t* ptr, uint16_t len);
